+++ BackgroundExecution/Screens/iPhone/HomeScreen_iPhone.xib.cs
@@ -2,8 +2,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using System.Threading;
 
 namespace Example_BackgroundExecution.Screens.iPhone
@@ -45,7 +45,8 @@ namespace Example_BackgroundExecution.Screens.iPhone
 		public void DoSomething ()
 		{
 			// register our background task
-			int taskID = UIApplication.SharedApplication.BeginBackgroundTask ( () => { this.BackgroundTaskExpiring (); });
+            // TODO: Removed (Action) cast added by migration tool
+			int taskID = (int)UIApplication.SharedApplication.BeginBackgroundTask ( () => { this.BackgroundTaskExpiring (); });
 
 			Console.WriteLine ("Starting background task " + taskID.ToString ());
 			
@@ -55,7 +56,7 @@ namespace Example_BackgroundExecution.Screens.iPhone
 			Console.WriteLine ("Background task " + taskID.ToString () + " completed.");
 			
 			// mark our background task as complete
-			UIApplication.SharedApplication.EndBackgroundTask (taskID);
+			UIApplication.SharedApplication.EndBackgroundTask ((nint)taskID);
 		}
 		
 		public void BackgroundTaskExpiring ()
